refactor: 리덕스로 모든 기능 구현
