+++ src/containers/chatroom/ChatRoomMainContainer.tsx
@@ -1,0 +1,63 @@
+import { useSelector } from 'react-redux';
+import React from 'react';
+import { useRef, useEffect } from 'react';
+import styled from 'styled-components';
+import { IChat } from '../../types/index';
+import ChatItem from '../../components/ChatItem';
+
+type TChatRoomMainProps = {
+  partnerUserId: number;
+};
+
+function ChatRoomMainContainer({ partnerUserId }: TChatRoomMainProps) {
+  const chatsState = useSelector(({ chats }: any) => ({
+    data: chats.data,
+  }));
+
+  const scrollRef = useRef<HTMLDivElement>(null);
+  useEffect(() => {
+    scrollRef.current!.scrollTo({
+      top: scrollRef.current!.scrollHeight,
+      behavior: 'smooth',
+    });
+  }, [chatsState.data]);
+
+  return (
+    <ChatRoomMainContainerBlock ref={scrollRef}>
+      {chatsState.data.map(({ userId, msg, unixTime }: IChat, idx: number) => (
+        <ChatItem
+          key={idx}
+          partnerUserId={partnerUserId}
+          userId={userId}
+          msg={msg}
+          unixTime={unixTime}
+        />
+      ))}
+    </ChatRoomMainContainerBlock>
+  );
+}
+
+const ChatRoomMainContainerBlock = styled.div`
+  width: 100%;
+  height: 75%;
+  @media screen and (min-width: 768px) {
+    height: 450px;
+  }
+  overflow-y: auto;
+
+  display: flex;
+  flex-direction: column;
+
+  &::-webkit-scrollbar {
+    width: 20px;
+  }
+
+  &::-webkit-scrollbar-thumb {
+    background: #e2e2e2;
+    background-clip: padding-box;
+    border-radius: 20px;
+    border: 8px solid transparent;
+  }
+`;
+
+export default React.memo(ChatRoomMainContainer);
